perf(residual): i16 residual representation end-to-end — match C's int16_t residual buffers
Forward transforms now take &[i16] (C's int16_t input), residual producers
emit i16 directly, and the tx pipeline/pd0/txt/mds3 scratch buffers are i16 —
halving residual staging footprint and removing the i32-vs-i16 divergence
named in issue #25. Inverse transforms keep &[i32] (coeff input). bd10 paths
narrow once at the residual producer (|u16 diff| <= 1023 fits i16).

Measured on the 512x512 qp32 p6 gradient cell: byte-identical to C and to
the parent revision, instruction count flat (451.68M vs 451.63M — the win
is halved buffer footprint, invisible to callgrind on a compute-bound cell).

Gates: nextest 3916/3916, regression_spotcheck 145/145, inter_byte 110/110,
dsp txfm/residual parity 146/146.

+++ rust/crates/svtav1-dsp/src/bench.rs
@@ -104,7 +104,7 @@ pub fn bench_fwd_txfm_throughput() -> Vec<BenchResult> {
 
     // 4x4 DCT
     {
-        let input: Vec<i32> = (0..16).map(|i| i * 7 - 50).collect();
+        let input: Vec<i16> = (0..16).map(|i| i * 7 - 50).collect();
         let mut output = [0i32; 16];
         let iterations = 500_000u64;
         let start = std::time::Instant::now();
@@ -127,7 +127,7 @@ pub fn bench_fwd_txfm_throughput() -> Vec<BenchResult> {
 
     // 8x8 DCT
     {
-        let input: Vec<i32> = (0..64).map(|i| i * 3 - 100).collect();
+        let input: Vec<i16> = (0..64).map(|i| i * 3 - 100).collect();
         let mut output = [0i32; 64];
         let iterations = 200_000u64;
         let start = std::time::Instant::now();
